// src/app/modules/admin/g-admin/position/wallet-bag-owner/wallet.types.ts
export interface BankPagination {
    length: number;
    size: number;
    page: number;
    lastPage: number;
    startIndex: number;
    endIndex: number;
    pagination: {
        page: number,
        pageSize: number,
        pageCount: number,
        total: number
    }
}
// ...
export interface AssetType {

    id: string;
    name: string,
    asset_sizes: [
        {
            id: number,
            name: string,
            size: string,
        }
    ]
}
// ...
export interface DataBank {
    data: [
        {
            id: number,
            branch_id: number,
            department_id: number,
            position_id: number,
            name: string,
            permission_id: number,
            user_id: number,
            first_name: string,
            last_name: string,
            account_number: string,
            email: string,
            image: string,
            image_signature: string,
            status: string,
            create_by: string,
            update_by: string,
            created_at: Date,
            updated_at: Date,
            No: number,
            quality: number
        }
    ]
}